[REFACTOR] #8 - dev rebase 이후 정상 동작하지 않는 부분 수정 1. domain의 MenuModel 위치 변경 2. DetailMenuViewHolder 제거 3. data binding의 imageSrc를 imageUrl로 변경
