+++ app/src/main/java/com/seom/banchan/ui/home/sidedish/SideDishViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.seom.banchan.R
 import com.seom.banchan.domain.model.*
+import com.seom.banchan.domain.model.home.toHomeMenuGridModel
 import com.seom.banchan.domain.usecase.GetSideMenusUseCase
 import com.seom.banchan.domain.usecase.GetSoupMenusUseCase
 import com.seom.banchan.ui.model.Model
