+++ app/src/main/java/com/seom/banchan/ui/model/CellType.kt
@@ -11,6 +11,5 @@ enum class CellType {
 
     IMAGE_SLIDER_CELL,
     IMAGE_LIST_CELL,
-    MENU_DETAIL_INFO_CELL,
-    MENU_DETAIL_EXTRA_CELL
+    MENU_DETAIL_INFO_CELL
 }
