+++ app/src/main/java/com/seom/banchan/ui/model/detail/MenuDetailModel.kt
@@ -4,8 +4,7 @@ import com.seom.banchan.ui.model.CellType
 import com.seom.banchan.ui.model.Model
 
 data class MenuDetailModel(
-    override val id: String,
-    override val type: CellType = CellType.MENU_DETAIL_EXTRA_CELL,
+    val id: String,
     val images: List<String>?,
     val name: String,
     val salePrice: Int,
@@ -14,4 +13,4 @@ data class MenuDetailModel(
     val deliveryInfo: String?,
     val deliveryFee: String?,
     val detailImages: List<String>?
-) : Model(id, type)
+)
