+++ app/src/main/java/com/seom/banchan/ui/model/home/HomeMenuGridModel.kt
@@ -1,6 +1,6 @@
 package com.seom.banchan.ui.model.home
 
-import com.seom.banchan.domain.model.MenuModel
+import com.seom.banchan.domain.model.home.MenuModel
 import com.seom.banchan.ui.model.CellType
 import com.seom.banchan.ui.model.Model
