+++ app/src/main/res/layout/item_menu_grid.xml
@@ -22,7 +22,7 @@
             android:id="@+id/iv_menu_thumbnail"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            app:imageSrc="@{menu.menu.image}"
+            app:imageUrl="@{menu.menu.image}"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
